rag_files/requirements.txt
# Technology Content RAG System - Dependencies

# Core LLM & RAG
llama-index==0.9.45
llama-index-embeddings-openai==0.1.6
llama-index-vector-stores-pinecone==0.1.4
llama-index-llms-openai==0.1.5

# Vector Database
pinecone-client==3.0.2

# API & Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Content Extraction
youtube-transcript-api==0.6.2
feedparser==6.0.10
beautifulsoup4==4.12.2
trafilatura==1.6.6
requests==2.31.0
lxml==4.9.3

# Orchestration
apache-airflow==2.7.3
apache-airflow-providers-postgres==5.10.1
apache-airflow-providers-python==4.10.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

# Logging & Monitoring
python-logging-loki==0.3.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# Optional: Audio Transcription
# deepgram-sdk==3.0.0  # For podcast transcription
# openai-whisper==20231117  # For audio transcription

# Optional: Advanced Search
# rank-bm25==0.2.2  # For BM25 hybrid search

# Optional: Semantic Caching
# redis==5.0.1

# Optional: Monitoring
# prometheus-client==0.19.0
# opentelemetry-api==1.21.0
